+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.2-pre
-*
+* Set rejectUnauthorized default to false
 
 ### 0.0.1
 * Initial version, extracted from cliffano/bagofholding